fix(ReDeathmatch): fix round mode switch condition
Fixes an issue where the round mode switch condition was not working as intended in the ReDeathmatch script. The previous condition only checked if the `redm_modes_switch` variable was equal to "disable", but it should also exclude cases where `redm_modes_switch` is a non-zero value. This fix adds an additional check to exclude non-zero values of `redm_modes_switch`.